Split test cases in separate assignment files
